refactor request for permission to timer
