+++ app/src/main/java/app/stacq/plan/ui/timer/PostNotificationsDialogFragment.kt
@@ -1,4 +1,4 @@
-package app.stacq.plan.ui.task
+package app.stacq.plan.ui.timer
 
 import android.app.AlertDialog
 import android.app.Dialog
@@ -11,7 +11,8 @@ class PostNotificationsDialogFragment : DialogFragment() {
     override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
         AlertDialog.Builder(requireContext())
             .setMessage(getString(R.string.post_notifications))
-            .setPositiveButton(getString(R.string.ok)) { _, _ -> }
+            .setPositiveButton(getString(R.string.allow)) { _, _ -> }
+            .setNegativeButton(getString(R.string.deny))
             .create()
 
     companion object {
